Create abstract methods in parent shape classes.
