+++ coxeter/shape_classes/base_classes.py
@@ -7,6 +7,7 @@ parts of the code base that either require or return shapes.
 """
 
 from abc import ABC, abstractmethod
+import numpy as np
 
 
 class Shape(ABC):
@@ -18,15 +19,27 @@ class Shape(ABC):
         """:math:`(3, )` :class:`numpy.ndarray` of float: Get or set the centroid of the shape."""  # noqa: E501
         pass
 
-    @center.setter
-    @abstractmethod
-    def center(self, value):
-        pass
-
     @property
     def gsd_shape_spec(self):
         """dict: Get a :ref:`complete GSD specification <shapes>`."""  # noqa: D401
-        return {}
+        raise NotImplementedError
+
+    def is_inside(self, points):
+        """Determine whether points are contained in this shape.
+
+        .. note::
+
+            Points on the boundary of the shape will return :code:`True`.
+
+        Args:
+            points (:math:`(N, 3)` :class:`numpy.ndarray`):
+                The points to test.
+
+        Returns:
+            :math:`(N, )` :class:`numpy.ndarray`:
+                Boolean array indicating which points are contained in the shape.
+        """
+        raise NotImplementedError
 
 
 class Shape2D(Shape):
@@ -38,10 +51,44 @@ class Shape2D(Shape):
         """float: Get or set the area of the shape."""
         pass
 
-    @area.setter
+    @property
     @abstractmethod
-    def area(self, value):
+    def perimeter(self):
+        """float: Get the perimeter of the shape."""
         pass
+
+    @property
+    def planar_moments_inertia(self):
+        r"""list[float, float, float]: Get the planar and product moments of inertia.
+
+        Moments are computed with respect to the :math:`x` and :math:`y`
+        axes. In addition to the two planar moments, this property also
+        provides the product of inertia.
+
+        The `planar moments <https://en.wikipedia.org/wiki/Polar_moment_of_inertia>`__
+        and the
+        `product <https://en.wikipedia.org/wiki/Second_moment_of_area#Product_moment_of_area>`__
+        of inertia define the in-plane area distribution.
+        """  # noqa: E501
+        raise NotImplementedError
+
+    @property
+    def polar_moment_inertia(self):
+        """float: Get the polar moment of inertia.
+
+        The `polar moment of inertia <https://en.wikipedia.org/wiki/Polar_moment_of_inertia>`__
+        is always calculated about an axis perpendicular to the shape (i.e. the
+        normal vector).
+
+        The polar moment is computed as the sum of the two planar moments of inertia.
+        """  # noqa: E501
+        return np.sum(self.planar_moments_inertia[:2])
+
+    @property
+    def iq(self):
+        """float: The isoperimetric quotient."""
+        # TODO: Add reference to math and some resource in the docstring.
+        return 4 * np.pi * self.area / (self.perimeter ** 2)
 
 
 class Shape3D(Shape):
@@ -53,23 +100,18 @@ class Shape3D(Shape):
         """float: Get or set the volume of the shape."""
         pass
 
-    @volume.setter
-    @abstractmethod
-    def volume(self, value):
-        pass
-
     @property
     @abstractmethod
     def surface_area(self):
         """float: Get or set the surface area of the shape."""
         pass
 
-    @surface_area.setter
-    @abstractmethod
-    def surface_area(self, value):
-        pass
-
-    @abstractmethod
     def inertia_tensor(self):
         """:math:`(3, 3)` :class:`numpy.ndarray`: Get the inertia tensor."""
-        pass
+        raise NotImplementedError
+
+    @property
+    def iq(self):
+        """float: The isoperimetric quotient."""
+        # TODO: allow for non-spherical reference ratio (changes the prefactor)
+        return np.pi * 36 * self.volume ** 2 / (self.surface_area ** 3)
