+++ coxeter/shape_classes/ellipse.py
@@ -148,18 +148,6 @@ class Ellipse(Shape2D):
         return i_x, i_y, i_xy
 
     @property
-    def polar_moment_inertia(self):
-        """float: Get the polar moment of inertia.
-
-        The `polar moment of inertia <https://en.wikipedia.org/wiki/Polar_moment_of_inertia>`__
-        is always calculated about an axis perpendicular to the ellipse (i.e. the
-        normal vector) placed at the centroid of the ellipse.
-
-        The polar moment is computed as the sum of the two planar moments of inertia.
-        """  # noqa: E501
-        return np.sum(self.planar_moments_inertia[:2])
-
-    @property
     def iq(self):
         """float: The isoperimetric quotient."""
         return np.min([4 * np.pi * self.area / (self.perimeter ** 2), 1])
